# OpenAPI_Swagger.yaml
openapi: 3.0.3
info:
  title: Pet Store API
  description: API para la gestión de productos y servicios de una tienda de mascotas.
  version: 1.0.0
servers:
  - url: https://api.petstore.com/v1
...
paths:
...
  /api/users/register:
    post:
      tags:
        - users
      summary: Registrar un nuevo usuario
      description: Permite a un nuevo usuario registrarse en la tienda.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: Usuario registrado exitosamente
        '400':
          description: Datos de entrada inválidos
  /api/users/login:
    post:
      tags:
        - users
      summary: Iniciar sesión
      description: Permite a un usuario iniciar sesión en la tienda.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Sesión iniciada exitosamente
        '401':
          description: Credenciales inválidas
  /api/users/profile:
    get:
      tags:
        - users
      summary: Obtener perfil de usuario
      description: Permite obtener la información del perfil de usuario autenticado.
      operationId: getUserProfile
      responses:
        '200':
          description: Perfil de usuario obtenido exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Usuario no autenticado

    put:
      tags:
        - users
      summary: Actualizar perfil de usuario
      description: Permite a un usuario actualizar su perfil.
      operationId: updateUserProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfile'
      responses:
        '200':
          description: Perfil de usuario actualizado exitosamente
        '400':
          description: Datos de entrada inválidos
  /api/users/addresses:
    get:
      tags:
        - users
      summary: Listar direcciones de envío
      description: Permite a un usuario autenticado obtener la lista de sus direcciones de envío.
      operationId: listUserAddresses
      responses:
        '200':
          description: Lista de direcciones obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Address'
        '401':
          description: Usuario no autenticado

    post:
      tags:
        - users
      summary: Agregar dirección de envío
      description: Permite a un usuario agregar una nueva dirección de envío.
      operationId: addUserAddress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Address'
      responses:
        '201':
          description: Dirección de envío agregada exitosamente
        '400':
          description: Datos de entrada inválidos
        '401':
          description: Usuario no autenticado
...
components:
  schemas:
...
    User:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    UserProfile:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        address:
          type: string
    Address:
      type: object
      properties:
        id:
          type: string
          description: ID único de la dirección
        street:
          type: string
          description: Calle de la dirección
        city:
          type: string
          description: Ciudad de la dirección
        state:
          type: string
          description: Estado o provincia de la dirección
        postalCode:
          type: string
          description: Código postal de la dirección
        country:
          type: string
          description: País de la dirección
...